make scoped rw lock private
